e-coin-jkpark/requirements.txt
Django==1.8
psycopg2==2.7.1
PyYAML==3.12
redis==2.10.5
